pyproject.toml
[tool.poetry]
name = "argrilocal-backend"
version = "0.1.0"
description = ""
authors = ["Geon-Woo <moongunwoo7019@naver.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"
djangorestframework = "^3.15.2"
django-environ = "^0.11.2"
djangorestframework-simplejwt = "^5.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
